+++ ValheimFortress/ValheimFortress.cs
@@ -21,7 +21,7 @@ namespace ValheimFortress
     {
         public const string PluginGUID = "MidnightsFX.ValheimFortress";
         public const string PluginName = "ValheimFortress";
-        public const string PluginVersion = "0.23.4";
+        public const string PluginVersion = "0.30.0";
 
         public static AssetBundle EmbeddedResourceBundle;
         public VFConfig cfg;